Add ability to customize Models and Categories trees
