+++ packages/itwin/tree-widget/README.md
@@ -152,6 +152,50 @@ The Models tree can be used in a "focus mode" where the tree is automatically fi
 
 ![Models tree focus mode demo](./media/models-tree-focus-mode.gif)
 
+#### Custom models tree
+
+This package provides building blocks for custom models tree:
+
+- `useModelsTree` - hook for creating and managing models tree state.
+- `useModelsTreeButtonProps` - hook for creating props for models tree buttons.
+
+Example:
+
+```tsx
+function CustomModelsTreeComponent({ imodel, viewport, getSchemaContext, selectionStorage, height, width }: CustomModelsTreeProps) {
+  const buttonProps = useModelsTreeButtonProps({ imodel, viewport });
+  const { modelsTreeProps, rendererProps } = useModelsTree({ activeView: viewport });
+
+  return (
+    <TreeWithHeader buttons={[<ModelsTreeComponent.ShowAllButton {...buttonProps} />, <ModelsTreeComponent.HideAllButton {...buttonProps} />]}>
+      <VisibilityTree
+        {...modelsTreeProps}
+        getSchemaContext={getSchemaContext}
+        selectionStorage={selectionStorage}
+        imodel={imodel}
+        height={height}
+        width={width}
+        treeRenderer={(props) => <CustomModelsTreeRenderer {...props} {...rendererProps} />}
+      />
+    </TreeWithHeader>
+  );
+}
+
+type VisibilityTreeRendererProps = ComponentPropsWithoutRef<typeof VisibilityTreeRenderer>;
+type CustomModelsTreeRendererProps = Parameters<ComponentPropsWithoutRef<typeof VisibilityTree>["treeRenderer"]>[0];
+
+function CustomModelsTreeRenderer(props: CustomModelsTreeRendererProps) {
+  const getLabel = useCallback<Required<VisibilityTreeRendererProps>["getLabel"]>(
+    (node) => {
+      const originalLabel = props.getLabel(node);
+      return <>Custom node - {originalLabel}</>;
+    },
+    [props.getLabel],
+  );
+  return <VisibilityTreeRenderer {...props} getLabel={getLabel} getSublabel={getSublabel} />;
+}
+```
+
 ### Categories tree
 
 The component, based on the active view, renders a hierarchy of either spatial (3d) or drawing (2d) categories. The hierarchy consists of two levels - the category (spatial or drawing) and its sub-categories. There's also a header that renders categories search box and various visibility control buttons.
@@ -199,6 +243,56 @@ Available header buttons:
 - `ModelsTreeComponent.ShowAllButton` makes all categories and their subcategories displayed.
 - `ModelsTreeComponent.HideAllButton` makes all categories hidden.
 - `ModelsTreeComponent.InvertButton` inverts display of all categories.
+
+#### Custom categories tree
+
+This package provides building blocks for custom categories tree:
+
+- `useCategoriesTree` - hook for creating and managing categories tree state.
+- `useCategoriesTreeButtonProps` - hook for creating props for categories tree buttons.
+
+
+Example:
+
+```tsx
+function CustomCategoiresTreeComponent({ imodel, viewport, getSchemaContext, selectionStorage, height, width }: CustomCategoriesTreeProps) {
+  const buttonProps = useCategoriesTreeButtonProps({ imodel, viewport });
+  const { categoriesTreeProps, rendererProps } = useCategoriesTree({ activeView: viewport });
+
+  return (
+    <TreeWithHeader buttons={[<CategoriesTreeComponent.ShowAllButton {...buttonProps} />, <CategoriesTreeComponent.HideAllButton {...buttonProps} />]}>
+      <VisibilityTree
+        {...categoriesTreeProps}
+        getSchemaContext={getSchemaContext}
+        selectionStorage={selectionStorage}
+        imodel={imodel}
+        height={height}
+        width={width}
+        treeRenderer={(props) => <CustomCategoriesTreeRenderer {...props} {...rendererProps} />}
+      />
+    </TreeWithHeader>
+  );
+}
+
+type VisibilityTreeRendererProps = ComponentPropsWithoutRef<typeof VisibilityTreeRenderer>;
+type CustomCategoriesTreeRendererProps = Parameters<ComponentPropsWithoutRef<typeof VisibilityTree>["treeRenderer"]>[0];
+
+function CustomModelsTreeRenderer(props: CustomCategoriesTreeRendererProps) {
+  const getLabel = useCallback<Required<VisibilityTreeRendererProps>["getLabel"]>(
+    (node) => {
+      const originalLabel = props.getLabel(node);
+      return <>Custom node - {originalLabel}</>;
+    },
+    [props.getLabel],
+  );
+
+  const getSublabel = useCallback<Required<VisibilityTreeRendererProps>["getSublabel"]>(() => {
+    return <>Custom sub label</>;
+  }, []);
+
+  return <VisibilityTreeRenderer {...props} getLabel={getLabel} getSublabel={getSublabel} />;
+}
+```
 
 ### iModel content tree
