+++ packages/itwin/tree-widget/src/components/tree-header/TreeWithHeader.scss
