+++ packages/itwin/tree-widget/src/components/trees/index.ts
@@ -8,6 +8,12 @@ export { CategoriesTreeComponent } from "./categories-tree/CategoriesTreeCompone
 export { IModelContentTreeComponent } from "./imodel-content-tree/IModelContentTreeComponent";
 export { ExternalSourcesTreeComponent } from "./external-sources-tree/ExternalSourcesTreeComponent";
 
+export { useModelsTree } from "./models-tree/UseModelsTree";
+export { useCategoriesTree } from "./categories-tree/UseCategoriesTree";
+
+export { useModelsTreeButtonProps } from "./models-tree/ModelsTreeButtons";
+export { useCategoriesTreeButtonProps } from "./categories-tree/CategoriesTreeButtons";
+
 export { Tree } from "./common/components/Tree";
 export { VisibilityTree } from "./common/components/VisibilityTree";
 export { TreeRenderer } from "./common/components/TreeRenderer";
