+++ packages/itwin/tree-widget/src/tree-widget-react.ts
@@ -6,4 +6,5 @@
 export { TreeWidget } from "./TreeWidget";
 export { createTreeWidget, TreeWidgetComponent } from "./components/TreeWidgetUiItemsProvider";
 export { SelectableTree, SelectableTreeDefinition } from "./components/SelectableTree";
+export { TreeWithHeader } from "./components/tree-header/TreeWithHeader";
 export * from "./components/trees";
